feat(ui): remove void for alertList and WIP models
