+++ ui/src/styles/index.less
@@ -17,7 +17,7 @@ html {
   }
 }
 
-
+/* 
 .ant-select-item-option-selected:not(.ant-select-item-option-disabled) {
     background: var(--coo-primary-04);
 }
@@ -53,7 +53,7 @@ html {
 
 .cona-c-logo {
     border-bottom: 0;
-}
+} 
 
 .rdb-layout-secondary-column .l-secondary-column__header {
     border-bottom: 0;
@@ -108,25 +108,14 @@ html {
 
 .montserrat-headers {
     --coo-header-font:'Montserrat'
-}
+}*/
 
-a {
+/* a {
     color:var(--coo-dark);
 }
-
-
-
-/* .c-skeleton-input{
-
-    &.ant-skeleton-active .ant-skeleton-input::after{
-            background: linear-gradient(90deg, #c53838 25%, #d47b7b 37%, #a80808 63%);
-        }
-    
-}
-
  */
 
- .ant-skeleton-active .ant-skeleton-input {
+/*  .ant-skeleton-active .ant-skeleton-input {
     background:rgba(var(--coo-rgb-secondary-03), 0);
  }
 
@@ -138,9 +127,9 @@ a {
     background: linear-gradient(90deg, #c53838 25%, #d47b7b 37%, #a80808 63%);
     background: linear-gradient(90deg, rgba(var(--coo-rgb-secondary-03), 0.5) 25%, rgba(var(--coo-rgb-secondary-03), 1) 37%, rgba(var(--coo-rgb-secondary-03), 0.2) 63%);
     border-radius: var(--coo-border-radius-small);
-} 
+}  */
 
-.ant-transfer-list-content-item:not(.ant-transfer-list-content-item-disabled):hover {
+/* .ant-transfer-list-content-item:not(.ant-transfer-list-content-item-disabled):hover {
     background-color: var(--coo-secondary-03);
     color:var(--coo-dark);
     border-radius: var(--coo-border-radius-small);
@@ -153,4 +142,4 @@ a {
 .ant-transfer-list-content-item-checked {
     background-color: var(--coo-secondary-03);
     color:var(--coo-dark)
-}
+} */
